// zarlai/frontend/src/gen/zarl/v1/admin_pb.d.ts
export declare type ListToolProvidersRequest = Message<"zarl.v1.ListToolProvidersRequest"> & {
};

/**
 * Describes the message zarl.v1.ListToolProvidersRequest.
 * Use `create(ListToolProvidersRequestSchema)` to create a new message.
 */
export declare const ListToolProvidersRequestSchema: GenMessage<ListToolProvidersRequest>;
// ...
export declare type UpdateToolProviderRequest = Message<"zarl.v1.UpdateToolProviderRequest"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;

  /**
   * @generated from field: bool enabled = 2;
   */
  enabled: boolean;

  /**
   * @generated from field: string config = 3;
   */
  config: string;
};

/**
 * Describes the message zarl.v1.UpdateToolProviderRequest.
 * Use `create(UpdateToolProviderRequestSchema)` to create a new message.
 */
export declare const UpdateToolProviderRequestSchema: GenMessage<UpdateToolProviderRequest>;
// ...
export declare type CreateToolProviderRequest = Message<"zarl.v1.CreateToolProviderRequest"> & {
  /**
   * @generated from field: string name = 1;
   */
  name: string;

  /**
   * @generated from field: string type = 2;
   */
  type: string;

  /**
   * @generated from field: string config = 3;
   */
  config: string;
};

/**
 * Describes the message zarl.v1.CreateToolProviderRequest.
 * Use `create(CreateToolProviderRequestSchema)` to create a new message.
 */
export declare const CreateToolProviderRequestSchema: GenMessage<CreateToolProviderRequest>;
// ...
export declare type DeleteToolProviderRequest = Message<"zarl.v1.DeleteToolProviderRequest"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;
};

/**
 * Describes the message zarl.v1.DeleteToolProviderRequest.
 * Use `create(DeleteToolProviderRequestSchema)` to create a new message.
 */
export declare const DeleteToolProviderRequestSchema: GenMessage<DeleteToolProviderRequest>;
// ...
export declare type ListRegisteredToolsRequest = Message<"zarl.v1.ListRegisteredToolsRequest"> & {
};

/**
 * Describes the message zarl.v1.ListRegisteredToolsRequest.
 * Use `create(ListRegisteredToolsRequestSchema)` to create a new message.
 */
export declare const ListRegisteredToolsRequestSchema: GenMessage<ListRegisteredToolsRequest>;
// ...
export declare type UpdateToolDescriptionRequest = Message<"zarl.v1.UpdateToolDescriptionRequest"> & {
  /**
   * @generated from field: string name = 1;
   */
  name: string;

  /**
   * @generated from field: string description = 2;
   */
  description: string;
};

/**
 * Describes the message zarl.v1.UpdateToolDescriptionRequest.
 * Use `create(UpdateToolDescriptionRequestSchema)` to create a new message.
 */
export declare const UpdateToolDescriptionRequestSchema: GenMessage<UpdateToolDescriptionRequest>;
// ...
export declare type ResetToolDescriptionRequest = Message<"zarl.v1.ResetToolDescriptionRequest"> & {
  /**
   * @generated from field: string name = 1;
   */
  name: string;
};

/**
 * Describes the message zarl.v1.ResetToolDescriptionRequest.
 * Use `create(ResetToolDescriptionRequestSchema)` to create a new message.
 */
export declare const ResetToolDescriptionRequestSchema: GenMessage<ResetToolDescriptionRequest>;